div_floor is only defined for v1x. Change some std::endl for '\n'. Change some double blank lines to a single blank line (in the output files).
